KUDU-1933. consensus: Avoid and repair integer overflow in log index
We observed a crash on a long-running master server that looked like the
following:

  F0308 00:25:53.568773  7655 log_index.cc:171] Check failed: log_index > 0 (-2147483648 vs. 0)

It turns out that this was caused due to integer overflow on the log
index field. This patch fixes this type of truncation in a couple of
places (LogReader and MakeOpId()) and adds a couple of new tests that
fail without both of those fixes.

This patch also adds "repair" logic in log replay during tablet
bootstrap that "reverts" integer overflow when it is detected while
rewriting the log entry.

Finally, this patch includes some test helper fixes to avoid log index
integer truncation in future tests.

Reviewed-on: http://gerrit.cloudera.org:8080/6376
Tested-by: Kudu Jenkins

+++ src/kudu/consensus/consensus-test-util.h
@@ -57,10 +57,10 @@ namespace consensus {
 using log::Log;
 using strings::Substitute;
 
-inline gscoped_ptr<ReplicateMsg> CreateDummyReplicate(int term,
-                                                      int index,
+inline gscoped_ptr<ReplicateMsg> CreateDummyReplicate(int64_t term,
+                                                      int64_t index,
                                                       const Timestamp& timestamp,
-                                                      int payload_size) {
+                                                      int64_t payload_size) {
     gscoped_ptr<ReplicateMsg> msg(new ReplicateMsg);
     OpId* id = msg->mutable_id();
     id->set_term(term);
@@ -89,13 +89,13 @@ inline RaftPeerPB FakeRaftPeerPB(const std::string& uuid) {
 inline void AppendReplicateMessagesToQueue(
     PeerMessageQueue* queue,
     const scoped_refptr<server::Clock>& clock,
-    int first,
-    int count,
-    int payload_size = 0) {
+    int64_t first,
+    int64_t count,
+    int64_t payload_size = 0) {
 
-  for (int i = first; i < first + count; i++) {
-    int term = i / 7;
-    int index = i;
+  for (int64_t i = first; i < first + count; i++) {
+    int64_t term = i / 7;
+    int64_t index = i;
     CHECK_OK(queue->AppendOperation(make_scoped_refptr_replicate(
         CreateDummyReplicate(term, index, clock->Now(), payload_size).release())));
   }
